ProgettoMameiIoT-master/requirements.txt
flask
paho-mqtt
google-cloud-firestore
google
flask-login